DAH-2705, polish: share the obfuscation-table test helper, name the probe's SSH-context contract
